[WIP] Update to non-forked hyper
Next steps: instead of `Send`ing the ResponseFuture we should
instead `Send` the `hyper::Request` in Transaction
